chore: add sqladmin fetcher
Adds a SQLAdminFetcher class that handles GCP authentication using the
google-auth-library and @googleapis/sqladmin packages. This new class
should be responsible for performing the requests to the Cloud SQL
Admin APIs, for now only the method for retrieving instance metadata
is implemented (getInstanceMetadata) that reaches to the
/connectSettings endpoint to retrieve the ip address and CA
certificate info.

This changeset also adds the nock library as a dev dependency in order
to mock network calls in unit tests.