chore: refactor ip addresses
Moves the `IpAddresses` interface to its own module along with the
`parseIpAddresses` function that extracts the value we want to hold in
the connector out of the instance metadata response.

The goal is to make this interface easier to reuse across modules but
a side effect of this change is that we're also able to move unit
tests that check the parse ip addresses logic out of the `nock` based
tests used in `sqladmin-fetcher` which are slower to run.