feat: add Connector.getOptions method
Add the public API to the connector, exposing a `Connector` class that
has a `getOptions` instance method that returns an the configuration
object that users can use to start a new PostgreSQL (using the `pg`
driver) connection / Pool.

Also adds a `close` method to terminate all callbacks that are kept open
to refresh cloud sql instance data periodically.