chore: add cert parser
Removes previous logic that assumed that responses from SQL Admin
getEphemeralCertificate would contain an expiration time property.

The ephemeral certificate received from that API response is now parsed
in memory in order to extract the expiration time from the parsed value.

Note: The `@types/node` package had to be updated to point to a recent
version of the runtime types that contain the X509Certificate parser.