fix: silent refresh errors on active connection
When an instance already have active connections, it should not throw
errors when trying to refresh cloud instance certificates in the
background. Otherwise these errors will bubble up and stop the end user
application.

This changeset fixes it by adding a more resilient system to the
CloudSQLInstance refresh logic that silents errors occured during
refresh and keeps valid certificate data that can be used if a refresh
is still ongoing or if any error happens when during a refresh.

Fixes: https://github.com/GoogleCloudPlatform/cloud-sql-nodejs-connector/issues/201