chore: add port option to tls server test helper
While updating to the new version of our test framework, the tests would
start occasionally fail with an `EADDRINUSE` error from trying to
connect to an already in use port.

This changeset adds an option to the `setup-tls-server` helper module so
that we can use different ports in different files that may be run in
parallel.