refactor: CloudSqlInstance.forceRefresh() should schedule a refresh and return immediately.
This avoids the situation where a refresh process is forced to start but may never finish, causing the 
node process to hang indefiniately awaiting the refresh promise that does not resolve. This also adds
a new method used in the test cases: CloudSqlInstance.refreshComplete() which returns a promise that 
will resolve when the refresh operation has finished.